Move most Stripe Elements mocking into function
